pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "learning-os"
version = "1.0.0"
description = "An AI-native learning workspace scaffold — turns any directory into a structured, self-improving learning system powered by Agent Skills."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
keywords = ["learning", "ai", "cursor", "agent-skills", "education"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Education",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
learning-os = "learning_os.cli:main"

[project.urls]
Homepage = "https://github.com/husam-haque/learning-os"
Repository = "https://github.com/husam-haque/learning-os"
Issues = "https://github.com/husam-haque/learning-os/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/learning_os"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""
